AUTO-6263: requirements.txt: Add dependency version constraints based on Python version
Drop the version constraints from the dependencies listed in the
install_requires[] keyword argument to setup(). Instead, require that
"-r requirements.txt" be supplied to "pip install" and populate the file
with rules conditional on the version of Python being used. Determine
the rules for a dependency by examining the Python versions shown as
Requires and Classifiers on the documentation for its versions on PyPi.

Also, rearrange the list of dependencies in install_requires[] to be in
alphabetical order and indicate for which submodule the dependency is
for.